+++ lib/src/inline_parser.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library markdown.inline_parser;
+library markdown.src.inline_parser;
 
 import 'ast.dart';
 import 'document.dart';
@@ -26,7 +26,6 @@ class InlineParser {
     new TextSyntax(r'\s*[A-Za-z0-9]+'),
 
     // The real syntaxes.
-
     new AutolinkSyntax(),
     new LinkSyntax(),
     new ImageLinkSyntax(),
@@ -125,16 +124,18 @@ class InlineParser {
     _stack.add(new TagState(0, 0, null));
 
     while (!isDone) {
-      bool matched = false;
+      var matched = false;
 
       // See if any of the current tags on the stack match. We don't allow tags
-      // of the same kind to nest, so this takes priority over other possible // matches.
-      for (int i = _stack.length - 1; i > 0; i--) {
+      // of the same kind to nest, so this takes priority over other possible
+      // matches.
+      for (var i = _stack.length - 1; i > 0; i--) {
         if (_stack[i].tryMatch(this)) {
           matched = true;
           break;
         }
       }
+
       if (matched) continue;
 
       // See if the current text matches any defined markdown syntax.
@@ -144,6 +145,7 @@ class InlineParser {
           break;
         }
       }
+
       if (matched) continue;
 
       // If we got here, it's just text.
@@ -163,17 +165,17 @@ class InlineParser {
   }
 
   void writeTextRange(int start, int end) {
-    if (end > start) {
-      final text = source.substring(start, end);
-      final nodes = _stack.last.children;
+    if (end <= start) return;
 
-      // If the previous node is text too, just append.
-      if ((nodes.length > 0) && (nodes.last is Text)) {
-        final newNode = new Text('${nodes.last.text}$text');
-        nodes[nodes.length - 1] = newNode;
-      } else {
-        nodes.add(new Text(text));
-      }
+    var text = source.substring(start, end);
+    var nodes = _stack.last.children;
+
+    // If the previous node is text too, just append.
+    if (nodes.length > 0 && nodes.last is Text) {
+      var textNode = nodes.last as Text;
+      nodes[nodes.length - 1] = new Text('${textNode.text}$text');
+    } else {
+      nodes.add(new Text(text));
     }
   }
 
@@ -183,6 +185,7 @@ class InlineParser {
 
   // TODO(rnystrom): Only need this because RegExp doesn't let you start
   // searching from a given offset.
+  @deprecated
   String get currentSource => source.substring(pos, source.length);
 
   bool get isDone => pos == source.length;
@@ -210,19 +213,16 @@ abstract class InlineSyntax {
       // Write any existing plain text up to this point.
       parser.writeText();
 
-      if (onMatch(parser, startMatch)) {
-        parser.consume(startMatch[0].length);
-      }
+      if (onMatch(parser, startMatch)) parser.consume(startMatch[0].length);
       return true;
     }
+
     return false;
   }
 
   ///Test if this syntax matches the current source.
-  Match matches(InlineParser parser) {
-    final Match startMatch = pattern.firstMatch(parser.currentSource);
-    return startMatch != null && startMatch.start == 0 ? startMatch: null;
-  }
+  Match matches(InlineParser parser)
+  => pattern.matchAsPrefix(parser.source, parser.pos);
 
   bool onMatch(InlineParser parser, Match match);
 }
@@ -230,9 +230,10 @@ abstract class InlineSyntax {
 /// Matches stuff that should just be passed through as straight text.
 class TextSyntax extends InlineSyntax {
   final String substitute;
+
   TextSyntax(String pattern, {String sub})
-      : super(pattern),
-        substitute = sub;
+      : substitute = sub,
+        super(pattern);
 
   bool onMatch(InlineParser parser, Match match) {
     if (substitute == null) {
@@ -254,10 +255,9 @@ class AutolinkSyntax extends InlineSyntax {
   // TODO(rnystrom): Make case insensitive.
 
   bool onMatch(InlineParser parser, Match match) {
-    final url = match[1];
-
-    final anchor = new Element.text('a', escapeHtml(url))
-      ..attributes['href'] = url;
+    var url = match[1];
+    var anchor = new Element.text('a', escapeHtml(url));
+    anchor.attributes['href'] = url;
     parser.addNode(anchor);
 
     return true;
@@ -270,15 +270,13 @@ class TagSyntax extends InlineSyntax {
   final RegExp endPattern;
   final String tag;
 
-  TagSyntax(String pattern, {String tag, String end})
-      : super(pattern),
-        endPattern = new RegExp((end != null) ? end : pattern, multiLine: true),
-        tag = tag;
-  // TODO(rnystrom): Doing this.field doesn't seem to work with named args.
+  TagSyntax(String pattern, {this.tag, String end})
+      : endPattern = new RegExp((end != null) ? end : pattern, multiLine: true),
+        super(pattern);
 
   bool onMatch(InlineParser parser, Match match) {
-    parser._stack.add(
-        new TagState(parser.pos, parser.pos + match[0].length, this));
+    parser._stack
+        .add(new TagState(parser.pos, parser.pos + match[0].length, this));
     return true;
   }
 
@@ -292,16 +290,13 @@ class TagSyntax extends InlineSyntax {
 class LinkSyntax extends TagSyntax {
   final LinkResolver linkResolver;
 
-  /// Weather or not this link was resolved by a [Resolver]
-  bool resolved = false;
-
   /// The regex for the end of a link needs to handle both reference style and
   /// inline styles as well as optional titles for inline links. To make that
   /// a bit more palatable, this breaks it into pieces.
   static get linkPattern {
-    final refLink = r'\s?\[([^\]]*)\]'; // "[id]" reflink id.
-    final title = r'(?:[ ]*"([^"]+)"|)'; // Optional title in quotes.
-    final inlineLink = '\\s?\\(([^ )]+)$title\\)'; // "(url "title")" link.
+    var refLink = r'\s?\[([^\]]*)\]'; // "[id]" reflink id.
+    var title = r'(?:[ ]*"([^"]+)"|)'; // Optional title in quotes.
+    var inlineLink = '\\s?\\(([^ )]+)$title\\)'; // "(url "title")" link.
     return '\](?:($refLink|$inlineLink)|)';
 
     // The groups matched by this are:
@@ -322,19 +317,15 @@ class LinkSyntax extends TagSyntax {
     // resolver function ([linkResolver]) that may choose to handle
     // this. Otherwise, it's just treated as plain text.
     Link link;
-    if (isNullOrEmpty(match[1])) {
+    if (match[1] == null) {
       if (linkResolver == null) return null;
 
-      // Only allow implicit links if the content is just text.
-      // TODO(rnystrom): Do we want to relax this?
-      if (state.children.any((child) => child is! Text)) return null;
-      // If there are multiple children, but they are all text, send the
-      // combined text to linkResolver.
-      var textToResolve =
-          state.children.fold('', (oldVal, child) => oldVal + child.text);
+      // Treat the contents as unparsed text even if they happen to match. This
+      // way, we can handle things like [LINK_WITH_UNDERSCORES] as a link and
+      // not get confused by the emphasis.
+      var textToResolve = parser.source.substring(state.endPos, parser.pos);
 
       // See if we have a resolver that will generate a link for us.
-      resolved = true;
       final val = linkResolver(textToResolve, null);
       if (val == null || val is Node)
         return val;
@@ -350,16 +341,27 @@ class LinkSyntax extends TagSyntax {
       if (link == null) return null;
     }
 
-    final Element node = new Element('a', state.children)
-      ..attributes["href"] = escapeHtml(link.url)
-      ..attributes['title'] = escapeHtml(link.title);
+    return _createElement(parser, match, state);
+  }
 
-    cleanMap(node.attributes);
-    return node;
+  /// Given that [match] has matched both a title and URL, creates an `<a>`
+  /// [Element] for it.
+  Element _createElement(InlineParser parser, Match match, TagState state) {
+    var link = getLink(parser, match, state);
+    if (link == null) return null;
+
+    var element = new Element('a', state.children);
+
+    element.attributes["href"] = escapeHtml(link.url);
+    if (link.title != null) {
+      element.attributes['title'] = escapeHtml(link.title);
+    }
+
+    return element;
   }
 
   Link getLink(InlineParser parser, Match match, TagState state) {
-    if ((match[3] != null) && (match[3] != '')) {
+    if (match[3] != null && match[3] != '') {
       // Inline link like [foo](url).
       var url = match[3];
       var title = match[4];
@@ -377,10 +379,12 @@ class LinkSyntax extends TagSyntax {
     } else {
       var id;
       // Reference link like [foo] [bar].
-      if (match[2] == '')
-          // The id is empty ("[]") so infer it from the contents.
-          id = parser.source.substring(state.startPos + 1, parser.pos);
-      else id = match[2];
+      if (match[2] == '') {
+        // The id is empty ("[]") so infer it from the contents.
+        id = parser.source.substring(state.startPos + 1, parser.pos);
+      } else {
+        id = match[2];
+      }
 
       // References are case-insensitive.
       id = id.toLowerCase();
@@ -403,8 +407,9 @@ class LinkSyntax extends TagSyntax {
           (state.children[0] as Text).text, url);
 
   bool onMatchEnd(InlineParser parser, Match match, TagState state) {
-    Node node = createNode(parser, match, state);
+    var node = createNode(parser, match, state);
     if (node == null) return false;
+
     parser.addNode(node);
     return true;
   }
@@ -413,31 +418,29 @@ class LinkSyntax extends TagSyntax {
 /// Matches images like `![alternate text](url "optional title")` and
 /// `![alternate text][url reference]`.
 class ImageLinkSyntax extends LinkSyntax {
-  final LinkResolver linkResolver;
-  ImageLinkSyntax({this.linkResolver}) : super(pattern: r'!\[');
+  ImageLinkSyntax({LinkResolver linkResolver})
+      : super(linkResolver: linkResolver, pattern: r'!\[');
 
-  Node createNode(InlineParser parser, Match match, TagState state) {
-    var node = super.createNode(parser, match, state);
-    if (resolved) return node;
-    if (node == null) return null;
+  /// Creates an <a> element from the given complete [match].
+  Element _createElement(InlineParser parser, Match match, TagState state) {
+    var element = super._createElement(parser, match, state);
+    if (element == null) return null;
 
-    assert(node is Element);
-    final Element nd = node;
+    var image = new Element.withTag("img");
+    image.attributes["src"] = element.attributes["href"];
 
-    final Element imageElement = new Element.withTag("img")
-      ..attributes["src"] = node.attributes["href"]
-      ..attributes["title"] = node.attributes["title"]
-      ..attributes["alt"] = node.children
-      .map((e) => isNullOrEmpty(e) || e is! Text ? '' : e.text)
-      .join(' ');
+    if (element.attributes.containsKey("title")) {
+      image.attributes["title"] = element.attributes["title"];
+    }
 
-    cleanMap(imageElement.attributes);
+    var alt = element.children.map((e) => e is! Text ? "" : e.text).join(" ");
+    if (alt != "") image.attributes["alt"] = alt;
 
-    nd.children
+    element.children
       ..clear()
-      ..add(imageElement);
+      ..add(image);
 
-    return node;
+    return element;
   }
 }
 
@@ -471,8 +474,8 @@ class TagState {
   /// Attempts to close this tag by matching the current text against its end
   /// pattern.
   bool tryMatch(InlineParser parser) {
-    Match endMatch = syntax.endPattern.firstMatch(parser.currentSource);
-    if ((endMatch != null) && (endMatch.start == 0)) {
+    var endMatch = syntax.endPattern.matchAsPrefix(parser.source, parser.pos);
+    if (endMatch != null) {
       // Close the tag.
       close(parser, endMatch);
       return true;
@@ -489,14 +492,14 @@ class TagState {
     // means they are mismatched. Mismatched tags are treated as plain text in
     // markdown. So for each tag above this one, we write its start tag as text
     // and then adds its children to this one's children.
-    int index = parser._stack.indexOf(this);
+    var index = parser._stack.indexOf(this);
 
     // Remove the unmatched children.
-    final unmatchedTags = parser._stack.sublist(index + 1);
+    var unmatchedTags = parser._stack.sublist(index + 1);
     parser._stack.removeRange(index + 1, parser._stack.length);
 
     // Flatten them out onto this tag.
-    for (final unmatched in unmatchedTags) {
+    for (var unmatched in unmatchedTags) {
       // Write the start tag as text.
       parser.writeTextRange(unmatched.startPos, unmatched.endPos);