+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-## v0.7.1+2
+## 0.8.0
+
+* Switch tests to use [test][] instead of [unittest][].
+* Remove (probably unused) `resolved` field from `LinkSyntax`.
+
+[test]: https://pub.dartlang.org/packages/test
+[unittest]: https://pub.dartlang.org/packages/unittest
+
+## 0.7.2
+
+* Allow resolving links that contain inline syntax (#42).
+
+## 0.7.1+3
+
+* Updated homepage.
+
+## 0.7.1+2
 
 * Formatted code.